backend/requirements.txt
channels[daphne]>=4.0.0,<5.0.0
channels-auth-token-middlewares>=1.1.0,<1.2.0
channels-redis>=4.2.0,<4.3.0
django>=4.2,<5.0
djangorestframework>=3.14.0,<3.15.0
djangorestframework-simplejwt>=5.3.1,<5.4.0
django-cors-headers>=4.3.1,<4.4.0
django-filter>=23.5,<24.0
drf-spectacular>=0.27.1,<0.28.0
psycopg2-binary>=2.9.9,<2.10.0
flake8>=7.0.0,<8.0.0
